+++ packages/block-editor/src/components/inserter/block-types-tab.native.js
@@ -13,6 +13,8 @@ import useBlockTypeImpressions from './hooks/use-block-type-impressions';
 
 const NON_BLOCK_CATEGORIES = [ 'reusable' ];
 
+const ALLOWED_EMBED_VARIATIONS = [ 'core/embed' ];
+
 function BlockTypesTab( { onSelect, rootClientId, listProps } ) {
 	const clipboardBlock = useClipboardBlock( rootClientId );
 
@@ -22,7 +24,14 @@ function BlockTypesTab( { onSelect, rootClientId, listProps } ) {
 
 			const allItems = getInserterItems( rootClientId );
 			const blockItems = allItems.filter(
-				( { category } ) => ! NON_BLOCK_CATEGORIES.includes( category )
+				( { id, category } ) =>
+					! NON_BLOCK_CATEGORIES.includes( category ) &&
+					// We don't want to show all possible embed variations
+					// as different blocks in the inserter. We'll only show a
+					// few popular ones.
+					( category !== 'embed' ||
+						( category === 'embed' &&
+							ALLOWED_EMBED_VARIATIONS.includes( id ) ) )
 			);
 
 			return {