+++ packages/react-native-editor/android/app/src/main/java/com/gutenberg/MainApplication.java
@@ -34,6 +34,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
+import com.reactnativecommunity.webview.RNCWebViewPackage;
 import com.swmansion.gesturehandler.react.RNGestureHandlerPackage;
 import com.swmansion.reanimated.ReanimatedPackage;
 import com.swmansion.rnscreens.RNScreensPackage;
@@ -283,6 +284,7 @@ public class MainApplication extends Application implements ReactApplication, Gu
                         new SafeAreaContextPackage(),
                         new RNScreensPackage(),
                         new RNPromptPackage(),
+                        new RNCWebViewPackage(),
                         mRnReactNativeGutenbergBridgePackage);
             }